+++ docs_snippets/lib/usages/widgets/switch.dart
@@ -12,7 +12,8 @@ final switchWidget = FSwitch(
   semanticsLabel: 'Enable notifications switch',
   // {@endcategory}
   // {@category "Core"}
-  style: const .delta(labelPadding: .value(.only(left: 8))),
+  style: const .delta(trailingLabelStyle: .delta(labelPadding: .value(.only(left: 8)))),
+  leadingLabel: false,
   enabled: true,
   value: false,
   onChange: (value) {},