+++ forui/test/src/widgets/radio_golden_test.dart
@@ -133,4 +133,26 @@ void main() {
       });
     }
   }
+
+  testWidgets('leading-label', (tester) async {
+    await tester.pumpWidget(
+      TestScaffold(
+        child: SizedBox(
+          width: 300,
+          child: FRadio(
+            leadingLabel: true,
+            label: const Text('Direct messages and mentions'),
+            description: const Text('Only send me direct messages and mentions.'),
+            error: const Text('An option must be selected.'),
+            value: true,
+          ),
+        ),
+      ),
+    );
+
+    await expectLater(
+      find.byType(TestScaffold),
+      matchesGoldenFile('radio/leading-label.png'),
+    );
+  });
 }