+++ forui/test/src/widgets/switch_golden_test.dart
@@ -6,6 +6,29 @@ import 'package:forui/forui.dart';
 import '../test_scaffold.dart';
 
 void main() {
+  testWidgets('leading-label', (tester) async {
+    await tester.pumpWidget(
+      TestScaffold(
+        child: SizedBox(
+          width: 300,
+          child: FSwitch(
+            leadingLabel: true,
+            label: const Text('Airplane Mode'),
+            description: const Text('Disable all wireless connections.'),
+            error: const Text('Please enable to continue.'),
+            semanticsLabel: 'Airplane Mode',
+            value: true,
+          ),
+        ),
+      ),
+    );
+
+    await expectLater(
+      find.byType(TestScaffold),
+      matchesGoldenFile('switch/leading-label.png'),
+    );
+  });
+
   for (final theme in TestScaffold.themes) {
     for (final (checked, value) in [('checked', true), ('unchecked', false)]) {
       testWidgets('${theme.name} - $checked - unfocused', (tester) async {